+++ apps/web/src/App.tsx
@@ -114,6 +114,7 @@ import { CARMacrophageStudioPage } from './pages/CARMacrophageStudioPage'
 import { LNPEncapsulationStudioPage } from './pages/LNPEncapsulationStudioPage'
 import { SpatialGNNStudioPage } from './pages/SpatialGNNStudioPage'
 import { SpatialMicrodissectionStudioPage } from './pages/SpatialMicrodissectionStudioPage'
+import { RNAThermodynamicsStudioPage } from './pages/RNAThermodynamicsStudioPage'
 import { Login } from './pages/Login'
 import { Register } from './pages/Register'
 import { WorkspaceProvider } from './context/WorkspaceContext'
@@ -303,6 +304,7 @@ function App() {
                       <Route path="/smfret-kinetics" element={<smFRETKineticsStudioPage />} />
                       <Route path="/milestone-v1-9" element={<MilestoneV19StudioPage />} />
                       <Route path="/spatial-microdissection" element={<SpatialMicrodissectionStudioPage />} />
+                      <Route path="/rna-thermodynamics" element={<RNAThermodynamicsStudioPage />} />
           </Routes>
     </WorkspaceProvider>
   )