+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ## [v2.0.0] - 2026-09-24
 ### Added (Phases 162 - 168)
 - Phase 162: Autonomous Spatial Transcriptomics Microdissection & Subcellular Spot Deconvolution Engine
+- Phase 163: Autonomous Non-Coding RNA Secondary Structure Thermodynamics & Minimum Free Energy Folding Matrix
 
 ## [v1.9.0] - 2026-09-24
 ### Added (Phases 155 - 161)